Refactor to use `Double` for `quantityCreativity` and `temperature` fields, add email sharing functionality, PDF generation for sharing messages, and enrich exception handling. (#8)

+++ src/main/java/com/kaust/ms/manager/prompt/chat/domain/models/requests/ModelRequest.java
@@ -26,6 +26,6 @@ public class ModelRequest {
     private String model;
     /** quantityCreativity. */
     @Schema(description = "quantityCreativity",requiredMode = Schema.RequiredMode.REQUIRED)
-    private Integer quantityCreativity;
+    private Double quantityCreativity;
 
 }
